Removed premature optimization from ssh-agent.bash
function get_ssh_hash was prepared to accept arguments via stdin
and via normal args ($@) and inside scripts it was used only via piping
so I decided it's unnecessary to have logic for accepting arguments at
all. If there will be need we can go back to the history and revert the
changes.

Signed-off-by: Bart Smykla <bartek@smykla.com>